+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-19T07:57:33.807Z",
+    "2026-05-19T10:21:55.585Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {