+++ packages/orq-rc/RELEASES.md
@@ -9289,3 +9289,13 @@ Based on:
 - [typescript v4.10.0-rc.14] packages/orq-rc
 ### Releases
 - [NPM v4.10.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.14 - packages/orq-rc
+
+## 2026-05-19 10:24:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.763.1 (2.884.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.10.0-rc.16] packages/orq-rc
+### Releases
+- [NPM v4.10.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/4.10.0-rc.16 - packages/orq-rc